+++ server/models/session.js
@@ -1,7 +1,10 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const session = sequelize.define('session', {
-    // Nothing
+    complete: {
+      type: DataTypes.BOOLEAN,
+      defaultValue: false
+    }
   }, {});
   session.associate = function(models) {
     session.hasMany(models.activity, {